+++ pyinterpolate/kriging/models/block/area_to_area_poisson_kriging.py
@@ -5,11 +5,8 @@ Authors
 -------
 1. Szymon Moliński | @SimonMolinsky
 
-TODO
-----
-* log errors
-* control negative predictions and errors
 """
+import logging
 from typing import Dict, Union
 
 import geopandas as gpd
@@ -31,7 +28,8 @@ def area_to_area_pk(semivariogram_model: TheoreticalVariogram,
                     unknown_block_point_support: np.ndarray,
                     number_of_neighbors: int,
                     raise_when_negative_prediction=True,
-                    raise_when_negative_error=True):
+                    raise_when_negative_error=True,
+                    log_process=True):
     """
     Function predicts areal value in a unknown location based on the area-to-area Poisson Kriging
 
@@ -69,6 +67,9 @@ def area_to_area_pk(semivariogram_model: TheoreticalVariogram,
     raise_when_negative_error : bool, default=True
         Raise error when prediction error is negative.
 
+    log_process : bool, default=True
+        Log process info and debug info.
+
     Returns
     -------
     results : List
@@ -87,11 +88,11 @@ def area_to_area_pk(semivariogram_model: TheoreticalVariogram,
     if isinstance(point_support, Dict):
         dps = point_support
     else:
+        if log_process:
+            logging.info('Point support is transformed to dictionary')
         dps = transform_ps_to_dict(point_support)
 
     # Check ids
-
-
     kriging_data = select_poisson_kriging_data(
         u_block_centroid=unknown_block,
         u_point_support=unknown_block_point_support,
@@ -143,16 +144,13 @@ def area_to_area_pk(semivariogram_model: TheoreticalVariogram,
     try:
         w = np.linalg.solve(weights, k_ones)
     except TypeError:
+        if log_process:
+            logging.debug('Wrong dtypes used for np.linalg.solve, casting to float.')
         weights = weights.astype(np.float)
         k_ones = k_ones.astype(np.float)
         w = np.linalg.solve(weights, k_ones)
 
     zhat = values.dot(w[:-1])
-
-    if raise_when_negative_prediction:
-        if zhat < 0:
-            raise ValueError(f'Predicted value is {zhat} and it should not be lower than 0. Check your sampling '
-                             f'grid, samples, number of neighbors or semivariogram model type.')
 
     # Calculate prediction error
 
@@ -160,6 +158,13 @@ def area_to_area_pk(semivariogram_model: TheoreticalVariogram,
         u_idx = unknown_block[0][0]
     else:
         u_idx = unknown_block[0]
+
+    if zhat < 0:
+        if log_process:
+            logging.debug(f'Prediction below 0 for area {u_idx}')
+        if raise_when_negative_prediction:
+            raise ValueError(f'Predicted value is {zhat} and it should not be lower than 0. Check your sampling '
+                             f'grid, samples, number of neighbors or semivariogram model type.')
 
     distances_within_unknown_block = get_distances_within_unknown(unknown_block_point_support)
 
@@ -172,6 +177,8 @@ def area_to_area_pk(semivariogram_model: TheoreticalVariogram,
     sigmasq = semivariance_within_unknown - sig_base
 
     if sigmasq < 0:
+        if log_process:
+            logging.debug(f'Variance Error below 0 for area {u_idx}')
         if raise_when_negative_error:
             raise ValueError(f'Predicted error value is {sigmasq} and it should not be lower than 0. '
                              f'Check your sampling grid, samples, number of neighbors or semivariogram model type.')