+++ pyinterpolate/test/test_distance/test_angles_between_points.py
@@ -33,3 +33,22 @@ class TestCalcPoint2PointsAngle(unittest.TestCase):
         expected_angles = [90, 45, 0, 225]
         for idx, ang in enumerate(angles):
             self.assertEqual(ang, expected_angles[idx])
+
+    def test_4(self):
+        import pandas as pd
+
+        sample = 'samples/point_data/csv/meuse_grid.csv'
+        df = pd.read_csv(sample, usecols=['x', 'y'])
+        vec_b = df.values
+        angles = calc_angles(vec_b)
+
+        # This is wrong, other point should be chosen as an origin
+        all_equal = np.allclose(angles, 61.5, rtol=0.1)
+        self.assertTrue(all_equal)
+
+        # Now we choose origin in the middle of a set
+        origin = np.mean(vec_b, axis=0)
+        new_angles = calc_angles(vec_b, origin=origin)
+
+        self.assertAlmostEqual(np.mean(new_angles), 164, places=1)
+        self.assertAlmostEqual(np.std(new_angles), 96.1, places=1)