+++ pyinterpolate/variogram/empirical/covariance.py
@@ -117,7 +117,7 @@ def directional_covariogram(points: np.array,
     tolerance : float, default=1
                 Value in range (0-1] to calculate semi-minor axis length of the search area. If tolerance is close
                 to 0 then points must be placed at a single line with beginning in the origin of coordinate system
-                and angle given by y axis and direction parameter.
+                and direction given by y axis and direction parameter.
                     * The major axis length == step_size,
                     * The minor axis size == tolerance * step_size.
                     * Tolerance == 1 creates the omnidirectional covariogram.
@@ -204,7 +204,7 @@ def calculate_covariance(points: np.array,
     tolerance : float, default=1
                 Value in range (0-1] to calculate semi-minor axis length of the search area. If tolerance is close
                 to 0 then points must be placed at a single line with beginning in the origin of coordinate system
-                and angle given by y axis and direction parameter.
+                and direction given by y axis and direction parameter.
                     * The major axis length == step_size,
                     * The minor axis size == tolerance * step_size.
                     * Tolerance == 1 creates the omnidirectional covariogram.
@@ -245,7 +245,7 @@ def calculate_covariance(points: np.array,
         temperature. It changes from equator to poles, so in the N-S and S-N axes. The good idea is to test if
         our observations are correlated in a few different directions. The main difference between an omnidirectional
         covariogram and a directional covariogram is that we take into account a different subset of neighbors.
-        The selection depends on the angle (direction) of analysis. You may imagine it like that:
+        The selection depends on the direction (direction) of analysis. You may imagine it like that:
 
         - Omnidirectional covariogram: we test neighbors in a circle,
         - Directional covariogram: we test neighbors within an ellipse and one direction is major.