+++ pyinterpolate/viz/raster.py
@@ -93,7 +93,7 @@ def interpolate_raster(data,
 
     tolerance : float (in range [0, 1]), optional, default=1
         If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
-        the coordinate system and the angle given by y axis and direction parameter. If ``tolerance`` is ``> 0`` then
+        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0`` then
         the bin is selected as an elliptical area with major axis pointed in the same direction as the line
         for 0 tolerance: